% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-project.R
\name{create_pemr_project}
\alias{create_pemr_project}
\title{Create a new PEMr project}
\usage{
create_pemr_project(
  path = ".",
  aoi_name,
  aoi_file = NULL,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{aoi_name}{The name of your area of interest (AOI), will be used to
name the project.}

\item{aoi_file}{Optional path to a file (e.g., .shp or .gpkg) defining the project
area of interest}

\item{rstudio}{If \code{TRUE}, calls \code{\link[usethis:use_rstudio]{use_rstudio()}} to make the new package or
project into an \href{https://r-pkgs.org/workflow101.html#sec-workflow101-rstudio-projects}{RStudio Project}.
If \code{FALSE} and a non-package project, a sentinel \code{.here} file is placed so
that the directory can be recognized as a project by the
\href{https://here.r-lib.org}{here} or
\href{https://rprojroot.r-lib.org}{rprojroot} packages.}

\item{open}{If \code{TRUE}, \link[usethis:proj_activate]{activates} the new project:
\itemize{
\item If using RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}
}
\value{
path to the newly created project
}
\description{
It will create a new project directory inside the directory specified in
\code{path}, which will be created if necessary. By default opens as a new project
in a new RStudio window.
}
\details{
All of the default directories for a PEM project will be created, as will
template R files for the different stages of the analysis.

A small \code{fid.RDS} file will be created inside a \verb{_meta} folder, which holds
the directory structure. This file is read and consulted by many functions in
the PEMr family of packages for reading and writing files to/from their
default locations.

As much as possible it is recommended to not deviate from the structure set
up here, so that analyses are understandable and reproducible by
collaborators and from project to project. If there is something missing in
the default setup, please open an issue at
https://github.com/ninoxconsulting/PEMr/issues.
}
\examples{
\dontrun{
create_pemr_project()
}
}
