% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pemr.R
\docType{package}
\name{pemr-package}
\alias{pemr}
\alias{pemr-package}
\title{pemr: Use Data From Personal Environmental Monitors}
\description{
Read and manage data from files collected using a variety of personal environmental monitors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odeleongt/postr}
  \item Report bugs at \url{https://github.com/odeleongt/postr/issues/}
}

}
\author{
\strong{Maintainer}: Oscar de León \email{odeleon@ces.uvg.edu.gt}

}
\keyword{internal}
