% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pemr.R
\name{run_pemr}
\alias{run_pemr}
\title{Attach PEMr packages}
\usage{
run_pemr()
}
\value{
TRUE or FALSE
}
\description{
Calling \code{run_pemr()} will attach all of the PEMr family of packages: \code{PEMprepr},
\code{PEMsamplr}, and \code{PEMmodelr}.
}
\examples{
run_pemr()
}
\author{
Genevieve Perkins
}
