% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEMr-package.R
\docType{package}
\name{PEMr-package}
\alias{PEMr}
\alias{PEMr-package}
\title{PEMr: Meta package for development of machine-learning predictive ecosystem maps}
\description{
PEMr is a "meta" package that installs all of the packages in the PEMr family of packages. This is similar in spirit to how the "tidyverse" package operates as an umbrella for the packages in the "tidyverse" suite of packages.
}
\author{
\strong{Maintainer}: Genevieve Perkins \email{gperkins@ninoxconsulting.ca} (\href{https://orcid.org/0000-0003-2663-5904}{ORCID})

Authors:
\itemize{
  \item Will McKenzie \email{Will.MacKenzie@gov.bc.ca}
  \item Kiri Daust \email{kiri.daust@gov.bc.ca}
  \item Colin Chisholm \email{Colin.Chisholm@unbc.ca}
  \item Matthew Coghill \email{matthewcoghill@live.ca}
  \item Claire Armour \email{claire.armour@gov.bc.ca}
  \item Andy Teucher \email{andy@andyteucher.ca} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
