% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{data_frame}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{unnest}
\alias{reexports}
\alias{map2}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{purrr}{\code{\link[purrr]{map2}}}

  \item{tibble}{\code{\link[tibble]{data_frame}}}

  \item{tidyr}{\code{\link[tidyr]{unnest}}}
}}

