% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pemr.R
\name{run_pemr}
\alias{run_pemr}
\title{Run PEMr libraries}
\usage{
run_pemr()
}
\value{
TRUE or FALSE
}
\description{
Set up required libraries to complete Predictive Ecosystem Mapping. Checks first if the required libraires exist and downloads if needed.
}
\examples{
run_pemr()
}
\author{
Genevieve Perkins
}
